Pin playwright dependency to fixed version
We did not specify a fixed playwright version
which caused pipeline errors due to missing
browser when the package-lock.json was recreated.

The browsers were missing in the docker image,
because the docker image contained older versions
of the browsers.

To prevent an accidental playwright update, the
version is now pinned to a specific version and
can be updated when updating the docker image to
ensure the browser versions in the docker image
and required browser versions by playwright stay
in sync.